fixed and refractored files from mousui to newselection
