+++ src/main/java/interface_adapter/newselection/SelectionPresenter.java
@@ -1,7 +1,7 @@
 package interface_adapter.newselection;
 
 import use_case.newselection.NewSelectionOutputBoundary;
-import use_case.mouseui.SelectOutputData;
+import use_case.newselection.NewSelectionOutputData;
 
 import java.awt.*;
 import java.awt.image.BufferedImage;
@@ -13,37 +13,37 @@ public class SelectionPresenter implements NewSelectionOutputBoundary {
     }
 
     @Override
-    public void setSelectionImage(SelectOutputData selectOutputData) {
+    public void setSelectionImage(NewSelectionOutputData selectOutputData) {
         BufferedImage image = selectOutputData.getSelectionImage();
         selectionViewModel.setBuffedImage(image);
     }
 
     @Override
-    public void setSelectionBounds(SelectOutputData selectOutputData) {
+    public void setSelectionBounds(NewSelectionOutputData selectOutputData) {
         Rectangle bounds = selectOutputData.getSelectionBounds();
         selectionViewModel.setSelectionBounds(bounds);
     }
 
     @Override
-    public void setIsDraggingSelection(SelectOutputData selectOutputData) {
-        boolean isDragging = selectOutputData.getIsDragginSelection();
+    public void setIsDraggingSelection(NewSelectionOutputData selectOutputData) {
+        boolean isDragging = selectOutputData.getIsDraggingSelection();
         selectionViewModel.setDraggingSelection(isDragging);
     }
 
     @Override
-    public void setHasSelection(SelectOutputData selectOutputData) {
+    public void setHasSelection(NewSelectionOutputData selectOutputData) {
         boolean hasSelection = selectOutputData.getHasSelection();
         selectionViewModel.setHasSelection(hasSelection);
     }
 
     @Override
-    public void setIsDrawing(SelectOutputData selectOutputData) {
+    public void setIsDrawing(NewSelectionOutputData selectOutputData) {
         boolean isDrawing = selectOutputData.getIsDrawing();
         selectionViewModel.setIsDrawing(isDrawing);
     }
 
     @Override
-    public void setSelectionToolBounds(SelectOutputData selectOutputData) {
+    public void setSelectionToolBounds(NewSelectionOutputData selectOutputData) {
         Rectangle bounds = selectOutputData.getSelectionToolBounds();
         selectionViewModel.setSelectionToolBounds(bounds);
     }
