+++ src/main/java/use_case/newselection/NewSelectionOutputBoundary.java
@@ -1,7 +1,10 @@
 package use_case.newselection;
 
-import use_case.mouseui.SelectOutputBoundary;
-
-public interface NewSelectionOutputBoundary extends SelectOutputBoundary{
-
+public interface NewSelectionOutputBoundary {
+    void setSelectionImage(NewSelectionOutputData selectOutputData);
+    void setSelectionBounds(NewSelectionOutputData selectOutputData);
+    void setIsDraggingSelection(NewSelectionOutputData selectOutputData);
+    void setHasSelection(NewSelectionOutputData selectOutputData);
+    void setIsDrawing(NewSelectionOutputData selectOutputData);
+    void setSelectionToolBounds(NewSelectionOutputData selectOutputData);
 }
